requirements.txt
customtkinter==5.2.2
darkdetect==0.8.0
packaging==24.1
psutil==6.0.0
py-cpuinfo==9.0.0
